Fix test for 32-bit environment

+++ pandas/tests/series/test_constructors.py
@@ -418,8 +418,8 @@ class TestSeriesConstructors(TestData):
         # GH 19342
         # construction with a numpy scalar
         # should not raise
-        result = Series(np.array(100), index=np.arange(4))
-        expected = Series(100, index=np.arange(4))
+        result = Series(np.array(100), index=np.arange(4), dtype='int64')
+        expected = Series(100, index=np.arange(4), dtype='int64')
         tm.assert_series_equal(result, expected)
 
     def test_constructor_corner(self):
